Raptor: rename raptor_init to raptor_solutions
